feat: cast int for queue_retry_after, define env file of .env.worker for worker

+++ processor-api/config/queue.php
@@ -63,7 +63,7 @@ return [
             'driver' => 'redis',
             'connection' => 'default',
             'queue' => env('REDIS_QUEUE', 'default'),
-            'retry_after' => env('QUEUE_RETRY_AFTER', 180),
+            'retry_after' => (int) env('QUEUE_RETRY_AFTER', 180),
             'block_for' => null,
         ],
